+++ app/src/main/java/com/space/quizapp/domain/usecase/points/insert/InsertQuizPointsUseCase.kt
@@ -1,5 +1,5 @@
 package com.space.quizapp.domain.usecase.points.insert
 
 interface InsertQuizPointsUseCase {
-    suspend fun invoke(userId: String, subjectId: Int, quizPoints: Int)
+    suspend fun invoke(userId: String, subjectId: Int, quizPoints: Double)
 }
